requirements.txt
pyautogui==0.9.54
pygetwindow==0.0.9
opencv-python==4.10.0.84
pytesseract==0.3.10
numpy==1.26.4
Pillow==10.3.0
